Backfill contract tests for storage
This change was inspired by the need to return the revision
number from update(). I realized at this point that I probably
should have been driving the behavior of InMemoryStorage with
contract tests all along, so that once I introduce FileStorage
I can be confident that it will have all the same behavior that
InMemoryStorage does.